refactor: Improve deletion of user conversations
This commit refactors the deleteAllUserConversations method in the conversation service to improve the deletion process. Instead of deleting all conversations where the user is a member, it now checks the number of members in each conversation. If a conversation has more than two members, the user is removed from the members list. Otherwise, the conversation is deleted entirely. This change enhances the handling of user conversations during deletion.